requirements.txt
streamlit
langchain-core
langchain-google-genai
langchain-openai
langchain-anthropic
google-generativeai
python-dotenv
fpdf
crawl4ai
pypdf
